Refactor client imports
