+++ examples/add_columns_to_table.py
@@ -5,7 +5,7 @@
 """
 
 from hive_metastore_client.builders.column_builder import ColumnBuilder
-from hive_metastore_client.hive_mestastore_client import HiveMetastoreClient
+from hive_metastore_client import HiveMetastoreClient
 
 HIVE_HOST = "<ADD_HIVE_HOST_HERE>"
 HIVE_PORT = 9083
