+++ examples/add_partitions.py
@@ -1,5 +1,5 @@
 from hive_metastore_client.builders import PartitionBuilder
-from hive_metastore_client.hive_mestastore_client import HiveMetastoreClient
+from hive_metastore_client import HiveMetastoreClient
 
 HIVE_HOST = "<ADD_HIVE_HOST_HERE>"
 HIVE_PORT = 9083
