+++ hive_metastore_client/__init__.py
@@ -1,1 +1,2 @@
 """Hive Metastore Client."""
+from hive_metastore_client.hive_mestastore_client import HiveMetastoreClient
