+++ tests/unit/hive_metastore_client/conftest.py
@@ -2,7 +2,7 @@ from unittest import mock
 
 import pytest
 
-from hive_metastore_client.hive_mestastore_client import HiveMetastoreClient
+from hive_metastore_client import HiveMetastoreClient
 
 
 @pytest.fixture
